working on enveloping tool

+++ pyNastran/utils/mathematics.py
@@ -53,8 +53,8 @@ def get_abs_max(min_values: np.ndarray,
                 max_values: np.ndarray,
                 dtype: str='float32') -> np.ndarray:
     """Get return the value with the greatest magnitude, preserving sign."""
-    min_values = np.asarray(min_values)
-    max_values = np.asarray(max_values)
+    min_values = np.asarray(min_values, dtype=dtype)
+    max_values = np.asarray(max_values, dtype=dtype)
     imin = np.abs(min_values) > np.abs(max_values)
     #out = np.zeros(min_values.shape, dtype=dtype)
     out = max_values.copy()
